+++ V2-Springboot/ALMXN/src/main/java/Project/ALMXN/Controllers/UsuarioController.java
@@ -62,7 +62,7 @@ public class UsuarioController {
     }
 
     @GetMapping("/editar")
-    public String mostrarEditar(@RequestParam("id") int idUsuario, Model model, HttpSession session) {
+    public String mostrarEditar(@RequestParam("id") Long idUsuario, Model model, HttpSession session) {
         Usuario usuario = (Usuario) session.getAttribute("usuarioLogueado");
         if (usuario == null) {
             return "redirect:/login";
@@ -74,7 +74,7 @@ public class UsuarioController {
     }
 
     @PostMapping("/eliminar")
-    public String eliminarUsuario(@RequestParam("idUsuario") Integer idUsuario, HttpSession session){
+    public String eliminarUsuario(@RequestParam("idUsuario") Long idUsuario, HttpSession session){
         Usuario usuarioLogueado = (Usuario) session.getAttribute("usuarioLogueado");
         if (usuarioLogueado == null || !usuarioLogueado.getRol().equals("Admin")) {
             return "redirect:/login";
@@ -89,7 +89,7 @@ public class UsuarioController {
     }
 
     @PostMapping("/activar")
-    public String activarProducto(@RequestParam("idUsuario") Integer idUsuario, HttpSession session) {
+    public String activarProducto(@RequestParam("idUsuario") Long idUsuario, HttpSession session) {
 
         Usuario usuario = (Usuario) session.getAttribute("usuarioLogueado");
         if (usuario == null || !usuario.getRol().equals("Admin")) {
